Initial draft of contao 4 compat
