+++ .travis.yml
@@ -1,15 +1,12 @@
 language: php
 
 php:
-  - "5.4"
-  - "5.5"
   - "5.6"
   - "7.0"
-  - "7.1"  
+  - "7.1"
 
 env:
-  - CONTAO_VERSION=~3.5.5
-
+  - CONTAO_VERSION=~4.4.0
 
 # Exclude impossible Contao Version combinations.
 matrix:
@@ -22,7 +19,7 @@ before_install:
 
 install:
   - travis_retry composer self-update && composer --version
-  - travis_retry composer require contao/core $CONTAO_VERSION --no-update
+  - travis_retry composer require contao/core-bundle $CONTAO_VERSION --no-update
   - travis_retry composer update --prefer-dist --no-interaction
 
 script: ant -keep-going
