+++ build.default.properties
@@ -7,3 +7,6 @@
 
 phpcs.standard=${basedir}/vendor/phpcq/coding-standard/phpcs/PhpCodeQuality/ruleset.xml
 phpmd.ruleset=${basedir}/vendor/phpcq/coding-standard/phpmd/ruleset.xml
+
+phpcs.excluded=src/CoreBundle/Resources/public \
+src/Resources/contao/languages \
