--- contao/config/autoload.ini
@@ -1,7 +1,0 @@
-
-;;
-; Configure what you want the autoload creator to register
-;;
-register_namespaces = false
-register_classes    = false
-register_templates  = true
