--- contao/config/autoload.php
@@ -1,31 +1,0 @@
-<?php
-
-/**
- * This file is part of MetaModels/attribute_tags.
- *
- * (c) 2012-2017 The MetaModels team.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- *
- * This project is provided in good faith and hope to be usable by anyone.
- *
- * @package    MetaModels
- * @subpackage AttributeTags
- * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @author     Andreas Isaak <info@andreas-isaak.de>
- * @author     Ingolf Steinhardt <info@e-spin.de>
- * @copyright  2012-2017 The MetaModels team.
- * @license    https://github.com/MetaModels/attribute_tags/blob/master/LICENSE LGPL-3.0
- * @filesource
- */
-
-/**
- * Register the templates
- */
-TemplateLoader::addFiles(
-    array
-    (
-        'mm_attr_tags' => 'system/modules/metamodelsattribute_tags/templates',
-    )
-);
