--- contao/config/event_listeners.php
@@ -1,43 +1,0 @@
-<?php
-
-/**
- * This file is part of MetaModels/attribute_tags.
- *
- * (c) 2012-2017 The MetaModels team.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- *
- * This project is provided in good faith and hope to be usable by anyone.
- *
- * @package    MetaModels
- * @subpackage AttributeTags
- * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @author     Ingolf Steinhardt <info@e-spin.de>
- * @copyright  2012-2017 The MetaModels team.
- * @license    https://github.com/MetaModels/attribute_tags/blob/master/LICENSE LGPL-3.0
- * @filesource
- */
-
-use MetaModels\Attribute\Events\CreateAttributeFactoryEvent;
-use MetaModels\Attribute\Tags\AttributeTypeFactory;
-use MetaModels\DcGeneral\Events\MetaModels\Tags\BackendSubscriber;
-use MetaModels\DcGeneral\Events\Table\Attribute\Tags\Subscriber;
-use MetaModels\Events\MetaModelsBootEvent;
-use MetaModels\MetaModelsEvents;
-
-return array
-(
-    MetaModelsEvents::SUBSYSTEM_BOOT_BACKEND => array(
-        function (MetaModelsBootEvent $event) {
-            new Subscriber($event->getServiceContainer());
-            new BackendSubscriber($event->getServiceContainer());
-        }
-    ),
-    MetaModelsEvents::ATTRIBUTE_FACTORY_CREATE => array(
-        function (CreateAttributeFactoryEvent $event) {
-            $factory = $event->getFactory();
-            $factory->addTypeFactory(new AttributeTypeFactory());
-        }
-    )
-);
