--- contao/html/.htaccess
@@ -1,7 +1,0 @@
-<IfModule mod_authz_core.c>
-    Require all granted
-</IfModule>
-<IfModule !mod_authz_core.c>
-    Order Deny,Allow
-    Allow from all
-</IfModule>
