+++ src/ContaoManager/Plugin.php
@@ -1,0 +1,49 @@
+<?php
+
+/**
+ * * This file is part of MetaModels/attribute_tags.
+ *
+ * (c) 2012-2017 The MetaModels team.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * This project is provided in good faith and hope to be usable by anyone.
+ *
+ * @package    MetaModels
+ * @subpackage AttributeTags
+ * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
+ * @copyright  2012-2017 The MetaModels team.
+ * @license    https://github.com/MetaModels/attribute_tags/blob/master/LICENSE LGPL-3.0
+ * @filesource
+ */
+
+namespace MetaModels\AttributeTagsBundle\ContaoManager;
+
+use Contao\ManagerPlugin\Bundle\BundlePluginInterface;
+use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
+use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
+use MetaModels\AttributeTagsBundle\MetaModelsAttributeTagsBundle;
+use MetaModels\CoreBundle\MetaModelsCoreBundle;
+
+/**
+ * Contao Manager plugin.
+ */
+class Plugin implements BundlePluginInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function getBundles(ParserInterface $parser)
+    {
+        return [
+            BundleConfig::create(MetaModelsAttributeTagsBundle::class)
+                ->setLoadAfter(
+                    [
+                        MetaModelsCoreBundle::class
+                    ]
+                )
+                ->setReplace(['metamodelsattribute_tags'])
+        ];
+    }
+}
