--- src/MetaModels/Attribute/Tags/AttributeTypeFactory.php
@@ -1,89 +1,0 @@
-<?php
-
-/**
- * This file is part of MetaModels/attribute_tags.
- *
- * (c) 2012-2017 The MetaModels team.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- *
- * This project is provided in good faith and hope to be usable by anyone.
- *
- * @package    MetaModels
- * @subpackage AttributeTags
- * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @author     Stefan Heimes <stefan_heimes@hotmail.com>
- * @author     Ingolf Steinhardt <info@e-spin.de>
- * @copyright  2012-2017 The MetaModels team.
- * @license    https://github.com/MetaModels/attribute_tags/blob/master/LICENSE LGPL-3.0
- * @filesource
- */
-
-namespace MetaModels\Attribute\Tags;
-
-use MetaModels\Attribute\IAttributeTypeFactory;
-
-/**
- * Attribute type factory for select attributes.
- */
-class AttributeTypeFactory implements IAttributeTypeFactory
-{
-    /**
-     * {@inheritdoc}
-     */
-    public function getTypeName()
-    {
-        return 'tags';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getTypeIcon()
-    {
-        return 'system/modules/metamodelsattribute_tags/html/tags.png';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function createInstance($information, $metaModel)
-    {
-        if (substr($information['tag_table'], 0, 3) === 'mm_') {
-            return new MetaModelTags($metaModel, $information);
-        }
-
-        return new Tags($metaModel, $information);
-    }
-
-    /**
-     * Check if the type is translated.
-     *
-     * @return bool
-     */
-    public function isTranslatedType()
-    {
-        return false;
-    }
-
-    /**
-     * Check if the type is of simple nature.
-     *
-     * @return bool
-     */
-    public function isSimpleType()
-    {
-        return false;
-    }
-
-    /**
-     * Check if the type is of complex nature.
-     *
-     * @return bool
-     */
-    public function isComplexType()
-    {
-        return true;
-    }
-}
