+++ src/Resources/config/services.yml
@@ -1,0 +1,10 @@
+services:
+    metamodels.attribute_tags.factory:
+        class: MetaModels\AttributeTagsBundle\Attribute\AttributeTypeFactory
+        arguments:
+            - '@database_connection'
+            - '@metamodels.table_manipulator'
+            - '@metamodels.factory'
+            - '@metamodels.filter_setting_factory'
+        tags:
+          - { name: metamodels.attribute_factory }
