+++ src/Resources/contao/languages/de/tl_metamodel_attribute.php
