--- tests/MetaModels/Test/Attribute/Tags/TagsAttributeTypeFactoryTest.php
@@ -1,130 +1,0 @@
-<?php
-
-/**
- * This file is part of MetaModels/attribute_tags.
- *
- * (c) 2012-2017 The MetaModels team.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- *
- * This project is provided in good faith and hope to be usable by anyone.
- *
- * @package    MetaModels
- * @subpackage AttributeTags
- * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @author     Ingolf Steinhardt <info@e-spin.de>
- * @copyright  2012-2017 The MetaModels team.
- * @license    https://github.com/MetaModels/attribute_tags/blob/master/LICENSE LGPL-3.0
- * @filesource
- */
-
-namespace MetaModels\Test\Attribute\Tags;
-
-use MetaModels\Attribute\IAttributeTypeFactory;
-use MetaModels\Attribute\Tags\AttributeTypeFactory;
-use MetaModels\IMetaModel;
-use MetaModels\Test\Attribute\AttributeTypeFactoryTest;
-
-/**
- * Test the attribute factory.
- */
-class TagsAttributeTypeFactoryTest extends AttributeTypeFactoryTest
-{
-    /**
-     * Mock a MetaModel.
-     *
-     * @param string $tableName        The table name.
-     *
-     * @param string $language         The language.
-     *
-     * @param string $fallbackLanguage The fallback language.
-     *
-     * @return IMetaModel
-     */
-    protected function mockMetaModel($tableName, $language, $fallbackLanguage)
-    {
-        $metaModel = $this->getMock(
-            'MetaModels\MetaModel',
-            array(),
-            array(array())
-        );
-
-        $metaModel
-            ->expects($this->any())
-            ->method('getTableName')
-            ->will($this->returnValue($tableName));
-
-        $metaModel
-            ->expects($this->any())
-            ->method('getActiveLanguage')
-            ->will($this->returnValue($language));
-
-        $metaModel
-            ->expects($this->any())
-            ->method('getFallbackLanguage')
-            ->will($this->returnValue($fallbackLanguage));
-
-        return $metaModel;
-    }
-
-    /**
-     * Override the method to run the tests on the attribute factories to be tested.
-     *
-     * @return IAttributeTypeFactory[]
-     */
-    protected function getAttributeFactories()
-    {
-        return array(new AttributeTypeFactory());
-    }
-
-    /**
-     * Test creation of an plain SQL tags attribute.
-     *
-     * @return void
-     */
-    public function testCreateTags()
-    {
-        $factory   = new AttributeTypeFactory();
-        $values    = array(
-            'tag_table'  => 'tl_page',
-            'tag_column' => 'pid',
-            'tag_alias'  => 'alias',
-        );
-        $attribute = $factory->createInstance(
-            $values,
-            $this->mockMetaModel('mm_test', 'de', 'en')
-        );
-
-        $this->assertInstanceOf('MetaModels\Attribute\Tags\Tags', $attribute);
-
-        foreach ($values as $key => $value) {
-            $this->assertEquals($value, $attribute->get($key), $key);
-        }
-    }
-
-    /**
-     * Test creation of an MetaModel referencing tags attribute.
-     *
-     * @return void
-     */
-    public function testCreateMetaModelTags()
-    {
-        $factory   = new AttributeTypeFactory();
-        $values    = array(
-            'tag_table'  => 'mm_page',
-            'tag_column' => 'pid',
-            'tag_alias'  => 'alias',
-        );
-        $attribute = $factory->createInstance(
-            $values,
-            $this->mockMetaModel('mm_test', 'de', 'en')
-        );
-
-        $this->assertInstanceOf('MetaModels\Attribute\Tags\MetaModelTags', $attribute);
-
-        foreach ($values as $key => $value) {
-            $this->assertEquals($value, $attribute->get($key), $key);
-        }
-    }
-}
