--- tests/bootstrap.php
@@ -1,38 +1,0 @@
-<?php
-
-/**
- * This file is part of MetaModels/attribute_tags.
- *
- * (c) 2012-2017 The MetaModels team.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- *
- * This project is provided in good faith and hope to be usable by anyone.
- *
- * @package    MetaModels
- * @subpackage AttributeTags
- * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @author     Ingolf Steinhardt <info@e-spin.de>
- * @copyright  2012-2017 The MetaModels team.
- * @license    https://github.com/MetaModels/attribute_tags/blob/master/LICENSE LGPL-3.0
- * @filesource
- */
-
-error_reporting(E_ALL);
-
-function includeIfExists($file)
-{
-	return file_exists($file) ? include $file : false;
-}
-
-if (
-	// Locally installed dependencies
-	(!$loader = includeIfExists(__DIR__.'/../vendor/autoload.php'))
-	// We are within an composer install.
-	&& (!$loader = includeIfExists(__DIR__.'/../../../autoload.php'))) {
-	echo 'You must set up the project dependencies, run the following commands:'.PHP_EOL.
-		'curl -sS https://getcomposer.org/installer | php'.PHP_EOL.
-		'php composer.phar install'.PHP_EOL;
-	exit(1);
-}
